Avoid externalizing the child app
Embroider is able to interoperate with the large amount of imports in a typical ember app that do not really come from modules is by deferring reslution to runtime if build-time resolution does not work.

That means when embroider sees an import from the package `child` and no such package is listed in the package.json dependencies, it assumes it's dealing with more runtime shenanigans and it shims the module with code that does `window.require('child/app')`, on the assumption that somebody somewhere is calling `window.define('child/app', ...)`.

This patch disables that behavior to get past the problem.

I also needed to set a publicPath on the child to get it loading its own assets.

The next failure revealed after these changes is that the child app is looking for its meta config tag and not finding it. That should hopefully be straightforward to fix by either creating a meta tag in the host app manually or by adjusting the child app to not use config-in-meta.